refactor: 장바구니 상품 삭제 시 id를 사용하도록 변경

+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModel.kt
@@ -28,9 +28,9 @@ class CartViewModel : ViewModel() {
         }
     }
 
-    fun deleteCartProduct(id: Int) {
+    fun deleteCartProduct(id: Long) {
         viewModelScope.launch {
-            cartRepository.deleteCartProduct(id.toLong())
+            cartRepository.deleteCartProduct(id)
             _onCartProductDeleted.value = true
         }
     }
